af-detect/requirements.txt
wfdb
numpy
scipy
pandas
torch
torchaudio
scikit-learn
tqdm
pyyaml